+++ tests/minimal_test/MinimalTest.cpp
@@ -22,7 +22,7 @@
 CPPUNIT_TEST_SUITE_REGISTRATION(MinimalTest);
 
 MinimalTest::MinimalTest() :
-test_dir("tests/test_files/")
+test_dir("../test_files/")
 {
     // no action
 }
@@ -32,7 +32,7 @@ MinimalTest::~MinimalTest() {
 }
 
 void MinimalTest::setUp() {
-    load_data_from("res/SAScore.dat");
+    load_data_from("../../res/SAScore.dat");
 }
 
 void MinimalTest::tearDown() {